+++ apps/console/src/features/claims/components/claims-list.tsx
@@ -1203,7 +1203,7 @@ export const ClaimsList: FunctionComponent<ClaimsListPropsInterface> = (
                 data-testid={ testId }
                 isRowSelectable={ (claim: Claim | ExternalClaim | ClaimDialect) =>
                     attributeConfig.isRowSelectable(claim) }
-                fixed
+                fixed={ attributeType !== OIDC }
             />
         </>
     );