+++ apps/console/src/features/connections/components/create/authenticator-create-wizard-factory.tsx
@@ -20,10 +20,12 @@ import { IdentifiableComponentInterface } from "@wso2is/core/models";
 import isEmpty from "lodash-es/isEmpty";
 import React, { FC, ReactElement, useEffect, useState } from "react";
 import { useTranslation } from "react-i18next";
+import { useSelector } from "react-redux";
 import { CreateConnectionWizard } from "./add-connection-wizard";
-import { 
-    EnterpriseConnectionCreateWizard 
+import {
+    EnterpriseConnectionCreateWizard
 } from "./enterprise-connection-create-wizard";
+import { AppState } from "../../../core";
 import { useGetConnectionTemplate, useGetConnections } from "../../api/connections";
 import { ConnectionManagementConstants } from "../../constants/connection-constants";
 import {
@@ -31,15 +33,15 @@ import {
     GenericConnectionCreateWizardPropsInterface,
     StrictConnectionInterface
 } from "../../models/connection";
-import { 
-    handleGetConnectionTemplateRequestError, 
-    handleGetConnectionsError 
+import {
+    handleGetConnectionTemplateRequestError,
+    handleGetConnectionsError
 } from "../../utils/connection-utils";
-import { 
-    ExpertModeAuthenticationProviderCreateWizard 
+import {
+    ExpertModeAuthenticationProviderCreateWizard
 } from "../wizards/expert-mode/expert-mode-authentication-provider-create-wizard";
-import { 
-    OrganizationEnterpriseConnectionCreateWizard 
+import {
+    OrganizationEnterpriseConnectionCreateWizard
 } from "../wizards/organization-enterprise/organization-enterprise-connection-create-wizard";
 import { TrustedTokenIssuerCreateWizard } from "../wizards/trusted-token-issuer-create-wizard";
 
@@ -112,11 +114,13 @@ export const AuthenticatorCreateWizardFactory: FC<AuthenticatorCreateWizardFacto
     ] = useState<ConnectionTemplateInterface>(undefined);
     const { t } = useTranslation();
 
+    const productName: string = useSelector((state: AppState) => state?.config?.ui?.productName);
+
     const {
         data: connectionsResponse,
         isLoading: isConnectionsFetchRequestLoading,
         error: connectionsFetchRequestError
-    } = useGetConnections(null, null, !selectedTemplate?.idp?.name 
+    } = useGetConnections(null, null, !selectedTemplate?.idp?.name
         ? "name sw " + selectedTemplate?.name : "name sw " + selectedTemplate?.idp?.name, null, true
     );
 
@@ -294,7 +298,7 @@ export const AuthenticatorCreateWizardFactory: FC<AuthenticatorCreateWizardFacto
                         title= { t("console:develop.features.authenticationProvider.templates.trustedTokenIssuer." +
                             "addWizard.title") }
                         subTitle= { t("console:develop.features.authenticationProvider.templates.trustedTokenIssuer." +
-                            "addWizard.subtitle") }
+                            "addWizard.subtitle", { productName }) }
                         onWizardClose={ () => {
                             setSelectedTemplateWithUniqueName(undefined);
                             setSelectedTemplate(undefined);
@@ -346,7 +350,7 @@ export const AuthenticatorCreateWizardFactory: FC<AuthenticatorCreateWizardFacto
                         { ...rest }
                     />
                 );
-                
+
             default:
                 return (
                     <CreateConnectionWizard