+++ modules/i18n/src/models/namespaces/console-ns.ts
@@ -2099,8 +2099,11 @@ export interface ConsoleNS {
                         group: {
                             heading: string;
                             hint: string;
+                            mappedRolesAbsentMessage: string;
+                            mappedRolesPresentMessage: string;
                             message: string;
                             placeHolder: string;
+                            roleMappingDisabledMessage: string;
                             label: string;
                             validation: {
                                 empty: string;